Refactor info.py into environment.py, fixes #5
