+++ HISTORY.rst
@@ -2,6 +2,10 @@
 History
 =======
 
+0.2.0 (2017-*)
+--------------
+* Refactor info.py into environment.py to account for regression in cpython micro release (#5)
+
 0.1.0 (2016-12-24)
 ------------------
