+++ pyenvdiff/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from .info import Environment, EnvironmentDiff  # noqa: F401
+from .environment import Environment, EnvironmentDiff  # noqa: F401
 from .version import __version__  # noqa: F401
 
 __author__ = """Jeffrey McLarty"""
